fix: change structure and fix routing
