+++ src/App.tsx
@@ -1,0 +1,12 @@
+import './App.css';
+import Results from './components/results/Results';
+import Search from './components/search/Search';
+
+export default function App() {
+  return (
+    <>
+      <Search />
+      <Results />
+    </>
+  );
+}
