+++ src/components/spiner/Spinner.tsx
@@ -1,0 +1,13 @@
+export default function Spinner() {
+  return (
+    <>
+      <img
+        width={60}
+        height={60}
+        src="/spinner.gif"
+        alt="loading"
+        style={{ margin: '20px' }}
+      />
+    </>
+  );
+}
