--- src/components/spinner.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import { ReactNode } from 'react';
-
-export class Spinner extends React.Component {
-  render(): ReactNode {
-    return (
-      <>
-        <img
-          width={60}
-          height={60}
-          src="/RSS_React/spinner.gif"
-          alt="loading"
-          style={{ margin: '20px' }}
-        />
-      </>
-    );
-  }
-}
