+++ vite.config.ts
@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  base: '/RSS_React/',
+  base: '/',
   server: {
     open: true,
   },
